Dev (#3)
* dev

* server auth msg encrypt

* edit

* port

* encrpt port data

* client control msg encrpted

* crypt done

* fix full buf bug

* make file fix

* client reconnect feature

* fix warning

* full buf fix

* fix stack over

* reconnect time set

* reset reconnect  time

* dir

* dir

* optimization

* optimization

* tell server local down

* dev

* dir

* dev

* dev

* dev

* readme

* bufsize

+++ xtun/src/net/epoll_demultiplexer.cpp
@@ -77,7 +77,7 @@ int EpollDemultiplexer::pollEvent(const EventHandlerMap &fileEvents,
     {
         return 0;
     }
-    if(firedEvents.capacity() < ret)
+    if(firedEvents.capacity() < static_cast<size_t>(ret))
     {
         firedEvents.resize(ret);
     }
